+++ source/DSCResources/DSC_xEnvironmentResource/DSC_xEnvironmentResource.psm1
@@ -8,8 +8,10 @@ Import-Module -Name (Join-Path -Path $modulePath `
     -ChildPath (Join-Path -Path 'xPSDesiredStateConfiguration.Common' `
         -ChildPath 'xPSDesiredStateConfiguration.Common.psm1'))
 
+Import-Module -Name (Join-Path -Path $modulePath -ChildPath 'DscResource.Common')
+
 # Import Localization Strings
-$script:localizedData = Get-LocalizedData -ResourceName 'DSC_xEnvironmentResource'
+$script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
 $script:envVarRegPathMachine = 'HKLM:\System\CurrentControlSet\Control\Session Manager\Environment'
 $script:envVarRegPathUser = 'HKCU:\Environment'