+++ source/DSCResources/DSC_xMsiPackage/DSC_xMsiPackage.psm1
@@ -11,8 +11,10 @@ Import-Module -Name (Join-Path -Path $modulePath `
     -ChildPath (Join-Path -Path 'xPSDesiredStateConfiguration.Common' `
         -ChildPath 'xPSDesiredStateConfiguration.Common.psm1'))
 
+Import-Module -Name (Join-Path -Path $modulePath -ChildPath 'DscResource.Common')
+
 # Import Localization Strings
-$script:localizedData = Get-LocalizedData -ResourceName 'DSC_xMsiPackage'
+$script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
 # Path to the directory where the files for a package from a file server will be downloaded to
 $script:packageCacheLocation = "$env:ProgramData\Microsoft\Windows\PowerShell\Configuration\BuiltinProvCache\MSFT_xMsiPackage"