+++ source/DSCResources/DSC_xWindowsPackageCab/DSC_xWindowsPackageCab.psm1
@@ -8,8 +8,10 @@ Import-Module -Name (Join-Path -Path $modulePath `
     -ChildPath (Join-Path -Path 'xPSDesiredStateConfiguration.Common' `
         -ChildPath 'xPSDesiredStateConfiguration.Common.psm1'))
 
+Import-Module -Name (Join-Path -Path $modulePath -ChildPath 'DscResource.Common')
+
 # Import Localization Strings
-$script:localizedData = Get-LocalizedData -ResourceName 'DSC_xWindowsPackageCab'
+$script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
 Import-Module -Name 'Dism'