+++ source/DSCResources/xFileUpload/xFileUpload.schema.psm1
@@ -8,6 +8,7 @@ Import-Module -Name (Join-Path -Path $modulePath `
     -ChildPath (Join-Path -Path 'xPSDesiredStateConfiguration.Common' `
         -ChildPath 'xPSDesiredStateConfiguration.Common.psm1'))
 
+Import-Module -Name (Join-Path -Path $modulePath -ChildPath 'DscResource.Common')
 <#
     .SYNOPSIS
         DSC Composite Resource uploads file or folder to an SMB share.