+++ source/DSCResources/xWindowsFeatureSet/xWindowsFeatureSet.schema.psm1
@@ -8,6 +8,7 @@ Import-Module -Name (Join-Path -Path $modulePath `
     -ChildPath (Join-Path -Path 'xPSDesiredStateConfiguration.Common' `
         -ChildPath 'xPSDesiredStateConfiguration.Common.psm1'))
 
+Import-Module -Name (Join-Path -Path $modulePath -ChildPath 'DscResource.Common')
 <#
     .SYNOPSIS
         A composite DSC resource to configure a set of similar xWindowsFeature resources.